// client/src/types/enums/Colors.ts
export enum Colors {
  none = '',
  white = '#FFF',
  light = '#F2F2F2',
  lightGrey = '#dcdcdc',
  grey = '#C4C4C4',
  greyDark = '#858585',
  greyDarker = '#7C7C7C',
  dark = '#151515',
  black = '#000',
  primary = '#640000',
  secondary = '#A5821E',
  green = '#32CA63',
  blue = '#1877f2',
  dev = '#206694',

  gryffindor = '#B11515',
  slytherin = '#067117',
  ravenclaw = '#3723B2',
  hufflepuff = 'E5C61E',

  danger = '#A10805',
  validation = '#05A136',

  transparent = 'transparent',

  stripe = '#5e56f2',
  paypal = '#012e85',
}